Implemented a confirm dialog

+++ src/com/ss/editor/ui/component/asset/tree/context/menu/action/DeleteFileAction.java
@@ -1,8 +1,11 @@
 package com.ss.editor.ui.component.asset.tree.context.menu.action;
 
+import com.ss.editor.JFXApplication;
 import com.ss.editor.Messages;
 import com.ss.editor.config.EditorConfig;
 import com.ss.editor.ui.component.asset.tree.resource.ResourceElement;
+import com.ss.editor.ui.dialog.ConfirmDialog;
+import com.ss.editor.ui.scene.EditorFXScene;
 
 import java.nio.file.Path;
 
@@ -10,14 +13,16 @@ import javafx.scene.control.MenuItem;
 import rlib.util.FileUtils;
 
 /**
- * Реализация действия по удалении файла.
+ * The action for deleting a file.
  *
- * @author Ronn
+ * @author JavaSaBr
  */
 public class DeleteFileAction extends MenuItem {
 
+    private static final JFXApplication JFX_APPLICATION = JFXApplication.getInstance();
+
     /**
-     * Элемент действия.
+     * The node in the tree.
      */
     private final ResourceElement element;
 
@@ -28,7 +33,7 @@ public class DeleteFileAction extends MenuItem {
     }
 
     /**
-     * Процесс удаления файла.
+     * Handle deleting.
      */
     private void processDelete() {
 
@@ -41,6 +46,19 @@ public class DeleteFileAction extends MenuItem {
             return;
         }
 
+        String question = Messages.ASSET_COMPONENT_RESOURCE_TREE_CONTEXT_MENU_DELETE_FILE_QUESTION;
+        question = question.replace("%file_name%", file.getFileName().toString());
+
+        final EditorFXScene scene = JFX_APPLICATION.getScene();
+        final ConfirmDialog confirmDialog = new ConfirmDialog(result -> handle(file, result), question);
+        confirmDialog.show(scene.getWindow());
+    }
+
+    /**
+     * Handle the answer.
+     */
+    private void handle(final Path file, final Boolean result) {
+        if (!result) return;
         FileUtils.delete(file);
     }
 }
